*If an user has SWS installed, these actions layouts will try to get original action name through SWS API functions instead of hardcoded.
I just think to tether SWS as required plug-in. This plug-in provides too many great function sometimes. But I'm remember self when i've crossovered to REAPER and didn't understand for  which reasons I have to install SWS, if all needed actions works for me without any add-ons. And this think stops me to use SWS right now instead any crutches. So, let it will be as little step to decision.