+Ability to set the position reset  when layout loads for properties also.
Now "remember position" option in configuration properties is adjustable:
"for not any" - neither sublayout nor properties position is not stores.
"Only for categories" - The position will be stored only for sublayouts. The properties position will be reset.
"Only for properties" - like as previous item, but it works for properties.
both for categories and properties" - All positions will be stored.