+Now a set method can return nil instead of empty string.
In this case, Properties Ribbon will output nothing to OSARA. It might be helpfull when either OSARA provides a report by itself and you will no needs to report an action or if a property shouldn't say anything, for example, when an user pressed Cancel button. Please note, that you still have to set custom undo label manualy.