SWS now should be installed
I was resist this all developing time, but finally the problems with avoiding of API extends which SWS provides reached my cup of patience. The main and big problem - is FX properties. In this realisation of, we too often should rely to FX names to avoid many assynchronous plug-ins definitions and working of exclude masks. Secondary, the macros for creating simple properties: SWS allows to get an action name and programmer shouldn't fill many strings to define its labels. Of course you still may fill it right now, but this labels now are optional. The other reasons are small and not important, but when we have less files the scripts complex works faster. For example, we are lost one module right now, because SWS provides its methods (for example byte words operations). Isn't it productive?
So, now the SWS extension should be installed for Properties Ribbon working. Maybe I'll demand it in specified cases at future, but for now, Properties Ribbon checks it when initialization has starts. Thanks for understanding!

Squashed commit of the following:

commit c18e13b8c5658244d0e7fb248fa9ee14b1e25d24
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Wed Jan 26 14:48:03 2022 +0300

    *Changed the caption of FX mask in Exclude parameters adding/editing

commit 38f098c46da7db37cf296ab5d25730c73fd036e9
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Wed Jan 26 01:18:37 2022 +0300

    *Changes of relying to static FX names

    Now all cases when FX properties demand static plugin names (exclude masks list, specific options per every parameter, assyncronously working plugin define) the real plug-in name used that as it stores on a FX distributive. Now an user can rename a FX in a chains - it no affect more to FX properties working.

    P.S.: MeldaProduction provides plug-in's libraries without author prefix, so I was forced to use the not precised pattern to define this. If any problems with MeldaProduction plug-ins wil occured, report me about and I'll add these plug-ins by independend part.

commit d0dcb7f09f3ac3ba8cc405ab6310cfe5b21e55d3
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Wed Jan 26 00:33:31 2022 +0300

    -Removed the pre-defined actions labels (where it's possible) because SWS can be make it for us.

commit 68fcc49e30b593f43f0ec7b8ee636f80c8c15bbe
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Wed Jan 26 00:25:10 2022 +0300

    *Changed the SWS presence checking method

commit a022d0f812d86cf11adbc7f6fa4109d8de70518c
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Wed Jan 26 00:18:05 2022 +0300

    *Changed the macros

    Now the label parameter is optional for each compose function. When no one value is passed, the get method will read the command name using SWS.

commit 4eaca733b26b5e1788c43840ea3e7fe893ee0b30
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Tue Jan 25 19:50:33 2022 +0300

    *Notices about SWS forced requiring

commit 85123ddbb211932245884ec1de6cb662ed2f6689
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Tue Jan 25 19:39:42 2022 +0300

    -Removed the bytewords module

    Now the byte words operations have getting through SWS extension methods.

commit 5bf1614dbd4ebf491967a37a56c57984d2367e6f
Author: Denis A. Shishkin <denis.outsidepro@gmail.com>
Date:   Tue Jan 25 19:21:50 2022 +0300

    *SWS extension now forcedly required.