+Parameter setting mode: added new action "Set midle parameter value"
This value is not precise because it not takes of values returned by REAPER like as it going on minimal and maximal parameters. There is no way to get it. It devides the maximal value by 2 then does minimal value plus given. Just please note that and not wait any cool result of.