Track navigation by category script has presented
Create needed category for your purposes then assign a track or selected tracks to appropriate category using extended interraction of a category. Then adjust this property to move the focus to the track which been assigned to this category. You have no keep all project matrix in your mind more!