Fixed ReaScript error in FX state switcher reporting
I don't know why there was attempt to get in FX index of itself, but not of parent object. Obviously there was another code earlier because it had work. And from which time there was this line... Sometimes I think that a otherwordly essences exist which write code with me together...