Implementation of new extended property for pan: Set sequential pan.
This property allows you to set the pan value for tracks/active takes of selected items using one panorama value and directions pattern. For example, you have 10 tracks which you want to pan to 50 percent left and right by pairs. In the first field specify 50 and in second field set the pattern "lr" or "left right". And all of ten tracks will be pan to the following order: 1 = 50% left, 2 = 50% right, 3 = 50% left, 4 = 50% right and etc.