composeThreePositionProperty: added optional parameter checkAbilityFunc
Allows to pass a function which should check the ability to do any changes for this object. Before the three-position property will attempt to do any changes, it will call this function and will pass either current object in a objects array or this object when it is not array. If the function returns true, the three-position property will change this object. If this function returns false, the property will no changes.
Syntax:
```Lua
---@param obj any
---@return boolean
function checkAbilityFunction(objItem)
```